퀘스트 클리어 / 스테이지 클리어/ 사망 시 팝업 ++ 방어력 적용수치 변경

+++ Assets/Scripts/IDamageable.cs
@@ -7,5 +7,5 @@ public interface IDamageable
     public BigInteger Health { get; set; }
     public bool Ondeath { get; set; }
 
-    public int Defence { get; set; }
+    public float Defence { get; set; }
 }
